Move simple structs into basic_structs.h
This was to avoid a circular dependency between the mhd and hydro util
files. Now this fundemental structs are on their own and can be
included as needed

Fix typo in Conserved struct
